+++ src/components/OverallMaturity/OverallMaturityOverview.tsx
@@ -2,8 +2,11 @@ import React from "react";
 import { InfoCard, Progress } from "@backstage/core-components";
 import Chart from "react-apexcharts";
 import { withTheme } from "@material-ui/core/styles";
-import { levelColorPalette } from "../helpers/level_color_helper";
-import { servicesByLevel, fontFamily } from "../helpers/maturity_report_helper";
+import { levelColorPalette } from "../../helpers/level_color_helper";
+import {
+  servicesByLevel,
+  fontFamily,
+} from "../../helpers/maturity_report_helper";
 
 type Props = {
   loading: Boolean;
